+++ townhall/myapi/tests/test_organization.py
@@ -1,0 +1,8 @@
+from django.test import TestCase
+from myapi import models as townhall_models
+from myapi import services as townhall_services
+
+# ORGANIZATION
+
+# Testing ALL tests "python manage.py test myapi.tests"
+# Testing ONLY organization "python manage.py test myapi.tests.test_organization"
